requirements.txt
numpy
pandas
matplotlib
seaborn
opencv-python-headless
torch
torchvision
scikit-learn
Pillow
tqdm
notebook
ipywidgets
fastapi
uvicorn[standard]
python-multipart
pydantic-settings